requirement.txt
google-generativeai
langchain_core
python-dotenv
streamlit
langchain_community
langserve